Move room related files from model module to shared module
